checks available for other modules
